get_child_points moved from users to points
