+++ zeton/data_access/points.py
@@ -21,3 +21,15 @@ def add_exp(exp, child_id):
     query = 'UPDATE main_points SET exp = exp + ?  WHERE child_id = ?;'
     get_db().execute(query, [exp, child_id])
     get_db().commit()
+
+def get_child_points(id):
+    query = """
+    SELECT mp.* 
+    FROM users AS u 
+    JOIN main_points AS mp on u.id = mp.child_id
+    WHERE u.id = ?
+    AND u.role = 'child'
+    """
+    result = get_db().execute(query, (id,))
+    child_points = dict(result.fetchone())
+    return child_points
